requirements.txt
sentence-transformers==2.7.0
faiss-cpu
numpy>=1.25,<2.0
pypdf
transformers==4.41.2
huggingface-hub==0.23.5
python-dotenv
langchain
langchain-groq
chromadb